Hotkey that can do calculations in the editor
After selecting something that looks like math eg. 1 + 5, hitting
<leader> then c will calculate the solution, and then replace the
selected math with the solution eg. 6.

This is done by piping the visual selection into bc -l, and then
replacing the visual selection with the output of bc -l.

bc -l, has strange syntax with regard to trig functions, such as sin is
now represented as s, so is cos. bc also does not have tan along with
most trigonometric and inverse trigonometric functions. The general
solution is to do the calculations using identities, but that obviously
is not reasonable.

Possible solutions to the problem include
1. awk/sed the input before putting it into bc.
2. Finding a different unix calculation command
3. Write own unix calculation command